boss upd8 everything on New Scene

+++ BULLET HELL/Assets/Scripts/Enemy/MiddleWare/Enemy_Shootpoint1Command.cs
@@ -8,7 +8,7 @@ public class Enemy_Shoot1Command : MonoBehaviour
     private Enemy_ShootingPattern shootingPattern;
     private BulletParticles particles;
 
-    public int desired_FireRate;
+    public float desired_FireRate;
     public float desired_BulletSpeed;
     public float desired_SpinSpeed;
     public float desired_Size;
